+++ pkg/volume/util/fs/fs_unsupported.go
@@ -16,7 +16,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package util
+package fs
 
 import (
 	"fmt"
